refactor(dnd): extract helpers and add docs
