+++ insight-fe/src/components/lesson/SlideSequencer.tsx
@@ -9,16 +9,15 @@ import { DropIndicator } from "@atlaskit/pragmatic-drag-and-drop-react-drop-indi
 import {
   draggable,
   dropTargetForElements,
-  monitorForElements,
 } from "@atlaskit/pragmatic-drag-and-drop/element/adapter";
 import { combine } from "@atlaskit/pragmatic-drag-and-drop/combine";
-import { reorder } from "@atlaskit/pragmatic-drag-and-drop/reorder";
 import {
   attachClosestEdge,
   extractClosestEdge,
   type Edge,
 } from "@atlaskit/pragmatic-drag-and-drop-hitbox/closest-edge";
 import { getReorderDestinationIndex } from "@atlaskit/pragmatic-drag-and-drop-hitbox/util/get-reorder-destination-index";
+import { useSlideDnD } from "@/hooks/useSlideDnD";
 
 export interface SlideBoard {
   id: string;
@@ -106,6 +105,7 @@ function SlideItem({
   const ref = useRef<HTMLDivElement | null>(null);
   const [closestEdge, setClosestEdge] = useState<Edge | null>(null);
 
+  // Initialize draggable slide item and track nearest edge for drop indication
   useEffect(() => {
     if (!ref.current) return;
     const el = ref.current;
@@ -192,68 +192,8 @@ export default function SlideSequencer({
     });
   }, [setSlides]);
 
-  useEffect(() => {
-    if (!containerRef.current) return;
-    return dropTargetForElements({
-      element: containerRef.current,
-      canDrop: ({ source }) =>
-        source.data.instanceId === instanceId.current &&
-        source.data.type === "slide",
-      getData: () => ({ columnId: "slides" }),
-      getIsSticky: () => true,
-    });
-  }, []);
-
-  useEffect(() => {
-    return monitorForElements({
-      canMonitor: ({ source }) => source.data.instanceId === instanceId.current,
-      onDrop: ({ source, location }) => {
-        if (source.data.type !== "slide") {
-          return;
-        }
-        if (!location.current.dropTargets.length) {
-          return;
-        }
-
-        const startIndex = slides.findIndex(
-          (s) => s.id === source.data.slideId,
-        );
-        if (startIndex === -1) return;
-
-        if (location.current.dropTargets.length === 1) {
-          const destinationIndex = getReorderDestinationIndex({
-            startIndex,
-            indexOfTarget: slides.length - 1,
-            closestEdgeOfTarget: null,
-            axis: "vertical",
-          });
-          setSlides((prev) =>
-            reorder({ list: prev, startIndex, finishIndex: destinationIndex }),
-          );
-          return;
-        }
-
-        if (location.current.dropTargets.length === 2) {
-          const [destinationRecord] = location.current.dropTargets;
-          const indexOfTarget = slides.findIndex(
-            (s) => s.id === destinationRecord.data.slideId,
-          );
-          const closestEdgeOfTarget = extractClosestEdge(
-            destinationRecord.data,
-          );
-          const destinationIndex = getReorderDestinationIndex({
-            startIndex,
-            indexOfTarget,
-            closestEdgeOfTarget,
-            axis: "vertical",
-          });
-          setSlides((prev) =>
-            reorder({ list: prev, startIndex, finishIndex: destinationIndex }),
-          );
-        }
-      },
-    });
-  }, [slides, setSlides]);
+  // Setup drop targets and reorder logic via dedicated hook
+  useSlideDnD(containerRef, slides, setSlides, instanceId);
 
   return (
     <Stack spacing={4}>
